dummy-repo/python-tool/requirements.txt
# Python Dependencies

# Web Framework
flask==1.0.2

# Database
mysql-connector-python==8.0.11
sqlite3

# YAML parsing
pyyaml==5.1

# HTTP client
requests==2.20.0

# Serialization
pickle5==0.0.9

# Cryptography
cryptography==2.3

# XML parsing
lxml==4.2.1

# Template engine
jinja2==2.10

# SQL toolkit
sqlalchemy==1.2.0

# Password hashing
passlib==1.7.1

# JWT handling
pyjwt==1.5.0

# Input validation
wtforms==2.2.0

# File uploads
werkzeug==0.14.1

# Markdown parsing
markdown==2.6.9

# Image processing
pillow==5.0.0

# PDF generation
reportlab==3.4.0

# XML processing
xml==1.0.0

# Testing
pytest==3.0.0
coverage==4.0

# Development
flask-debugtoolbar==0.10.0

# Utilities
urllib3==1.22
paramiko==2.0.0
